# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel Agency API",
        "description": "توثيق API لنظام إدارة وكالات السفر",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/agencies": {
            "get": {
                "tags": [
                    "Agencies"
                ],
                "summary": "عرض جميع الوكالات",
                "operationId": "697a4a7e63ba7c3bbce65c71a718e22e",
                "responses": {
                    "200": {
                        "description": "قائمة الوكالات"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agencies"
                ],
                "summary": "إضافة وكالة جديدة مع أدمن وكالة",
                "operationId": "a5d73eab2b5be570294a40eb452b9e99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "agency_name",
                                    "admin_name",
                                    "admin_email",
                                    "admin_password"
                                ],
                                "properties": {
                                    "agency_name": {
                                        "type": "string"
                                    },
                                    "admin_name": {
                                        "type": "string"
                                    },
                                    "admin_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "admin_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "license_number": {
                                        "type": "string"
                                    },
                                    "commercial_record": {
                                        "type": "string"
                                    },
                                    "tax_number": {
                                        "type": "string"
                                    },
                                    "logo": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "license_expiry_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "currency": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء الوكالة والأدمن بنجاح"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/agencies/{id}": {
            "get": {
                "tags": [
                    "Agencies"
                ],
                "summary": "عرض وكالة محددة",
                "operationId": "2a27ca9da6bab720adb9abca80c00512",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "بيانات الوكالة"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agencies"
                ],
                "summary": "تعديل بيانات وكالة",
                "operationId": "a65afa3ec5200081ee6fc208cc19bce5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم التحديث بنجاح"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agencies"
                ],
                "summary": "حذف وكالة",
                "operationId": "6fe4456f7d560b60ccae12c35231d03e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم الحذف بنجاح"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Agencies",
            "description": "إدارة الوكالات (للسوبر أدمن فقط)"
        }
    ]
}